refactor: Simplify notification navigation logic
This commit refactors the notification navigation logic within
`HomeActivity`. The `NotificationNavigation` class is removed and its
functionality is moved to a new `findNavigatingPath` extension
function.

The new logic uses `Uri.parse` directly and eliminates the need for
`attemptOrNull`. This simplifies the code and improves readability,
making it easier to understand and maintain. Also, passes the whole
intent to `ContactUsActivity` to handle the deep link.